Rimosse classi Vector inutili
